+++ src/sn/forage/dao/IVillage.java
@@ -8,7 +8,7 @@ import sn.forage.entities.Village;
 public interface IVillage {
 	public int add(Village village);
 	public List<Village> list();
-	public Village getVillage(Long id);
+	public Village getVillage(int id);
 	public Village update(Village c);
 	public void deleteVillage(Long id);
 }
